Removed garbage code, added comments.
Comments are now there to explain all the parts of the programs.

Removed some debugging garbage code - print statements and hard-codes.

Currently on master is my secret Wit.ai key... I've reset it and removed
it.

Please note that this branch isn't guaranteed to work yet - developing
on x64, but the full program can only run on Raspberry Pi because of the
compiled libraries. Haven't tested the whole thing. Next commit, I'll
remove the hardcoded paths in util.py.